Online Poll System Backend/requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
